Got to rename the variables to have underscores in so they don't appear as commands in !help
If you check blaskbot.py you'll notice that !help is automatically populated based on anything in commands.py that doesn't start with an underscore (which is why all the module imports do). New global variables should too. There's probably a better way to do this but I need to go live in 8 minutes so this is what we're doing right now.